Gestion de Imagenes finalizado

+++ app/Http/Controllers/TourController.php
@@ -9,74 +9,64 @@ use App\Models\TourPredefinido;
 use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-
+use Illuminate\Support\Str;
+use Validator;
 class TourController extends Controller
 {
-    public function InsertarTourArmado(Request $request)
-    {
-
-        try {
-            DB::beginTransaction();
-            $tour = TourArmado::create([
-                'usuarioId'      => $request->usuarioId,
-                'nombreTour'     => $request->nombreTour,
-                'horaInicioTour' => $request->horaInicioTour,
-            ]);
-
-            $puntosdeInteresTour = $request->puntosdeInteresTour;
-            $puntos              = explode(',', $puntosdeInteresTour);
-            $id                  = $tour->id;
-            foreach ($puntos as $punto) {
-                TourItems::create([
-                    'puntoInteresId' => $punto,
-                    'tourId'         => $id,
-                ]);
-            }
-            DB::commit();
-            return response()->json([
-                'Message' => 'Tour Creado correctamente ' . $tour->nombreTour,
-            ]);
-
-        } catch (\Exception$e) {
-            DB::rollBack();
-
-            return response()->json([
-                'Message' => 'No se pudo crear el tour, revise los datos enviados',
-            ]);
-        }
-
-    }
-
+    
     public function InsertarTourPredefinido(Request $request)
     {
-        try {
-            DB::beginTransaction();
-            $tour = TourPredefinido::create([
-                'nombreTourPredefinido'       => $request->nombreTourPredefinido,
-                'horaDeInicioTourPredefinido' => $request->horaDeInicioTourPredefinido,
-                'descripcionTourPredefinido'  => $request->descripcionTourPredefinido,
-            ]);
-
-            $puntosdeInteresTour = $request->puntosdeInteresTour;
-            $puntos              = explode(',', $puntosdeInteresTour);
-            $id                  = $tour->id;
-            foreach ($puntos as $punto) {
-                TourItemsPredefinido::create([
-                    'puntoInteresId' => $punto,
-                    'tourId'         => $id,
+        if($request->Opcion==='AltaDeTour'){
+            try {
+                DB::beginTransaction();
+                $tour = TourPredefinido::create([
+                    'nombreTourPredefinido'       => $request->nombreTourPredefinido,
+                    'horaDeInicioTourPredefinido' => $request->horaDeInicioTourPredefinido,
+                    'descripcionTourPredefinido'  => $request->descripcionTourPredefinido,
+                ]);
+    
+                $puntosdeInteresTour = $request->puntosdeInteresTour;
+                $puntos              = explode(',', $puntosdeInteresTour);
+                $id                  = $tour->id;
+                foreach ($puntos as $punto) {
+                    TourItemsPredefinido::create([
+                        'puntoInteresId' => $punto,
+                        'tourId'         => $id,
+                    ]);
+                }
+                DB::commit();
+                return response()->json([
+                    'Message' => 'Tour Predefinido Creado correctamente ' . $tour->nombreTourPredefinido,
+                ]);
+    
+            } catch (\Exception$e) {
+                DB::rollBack();
+    
+                return response()->json([
+                    'Message' => 'No se pudo crear el tour predefinido, revise los datos enviados',
                 ]);
             }
-            DB::commit();
-            return response()->json([
-                'Message' => 'Tour Predefinido Creado correctamente ' . $tour->nombreTourPredefinido,
-            ]);
-
-        } catch (\Exception$e) {
-            DB::rollBack();
-
-            return response()->json([
-                'Message' => 'No se pudo crear el tour predefinido, revise los datos enviados',
-            ]);
+        }
+        if($request->Opcion==='AltaDeImagenTour'){
+            if (!$request->hasFile('file')) {
+                return $this->returnError(202, 'file is required');
+            }
+            // return response()->json([
+            //         "codigo"    => "200",
+            //         "respuesta" => "Se agrego imagen con exito",
+            //         "idEvento" => $request->idTour
+            //     ]);
+            $response = Cloudinary::upload($request->file('file')->getRealPath(), ['folder' => 'feeluy']);
+                $url       = $response->getSecurePath();
+            
+                $tour=DB::table('tour_predefinido')
+                ->where('id','=',$request->idTour)
+                ->update(['imagen' => $url]);
+                return response()->json([
+                    "codigo"    => "200",
+                    "respuesta" => "Se agrego imagen con exito",
+                    "idTour" => $request->idTour
+                ]);
         }
 
     }
@@ -118,6 +108,10 @@ class TourController extends Controller
 
     public function ListarToursPredefinidos(Request $request)
     {
+        if($request->Opcion==='ImagenTour'){
+            $imagen = TourPredefinido::where('id','=',$request->tour_id)->get('imagen');
+            return response()->json($imagen);
+        }
         if($request->Opcion==='Unico'){
             $toursPredefinidos=TourPredefinido::with('TourItems.PuntosInteres')->get()->find($request->id);
             
@@ -142,31 +136,29 @@ class TourController extends Controller
         ]);
 
     }
-    public function destroy($id)
+    public function destroy(request $request,$id)
     {
-        $Tour=TourPredefinido::findOrFail($id);
-        $Tour->delete();
-         return response()->json([
-            "codigo"    => "200",
-            "respuesta" => "Se elimino con exito",
-            
-        ]);
+        if($request->Opcion==='EliminarImagen'){
+            $evento=DB::table('tour_predefinido')
+                ->where('id','=',$id)
+                ->update(['imagen'=>null]);
+             return response()->json([
+                "codigo"    => "200",
+                "respuesta" => "Se elimino con exito la Imagen",
+                "idTour" => $id
+            ]); 
+            }
+        if($request->Opcion==='EliminarTour'){
+            $Tour=TourPredefinido::findOrFail($id);
+            $Tour->delete();
+             return response()->json([
+                "codigo"    => "200",
+                "respuesta" => "Se elimino con exito",
+                
+            ]);
+        }
+        
     
     }
-    public function ModificarImagenesTour(Request $request,$idTour){
-        if (!$request->hasFile('file')) {
-            return $this->returnError(402, 'file is required');
-        }
-        $response = Cloudinary::upload($request->file('file')->getRealPath(), ['folder' => 'feeluy']);
-            $url       = $response->getSecurePath();
-        
-           
-        $Tour=TourPredefinido::findOrFail($idTour);
-        $Tour->imagen=$url;
-        $Tour->save();
-        return response()->headers('Access-Control-Allow-Origin',"*")->json([
-            "codigo"    => "200",
-            "respuesta" => "Se agrego imagen con exito",
-        ]);
-    }
+  
 }
